requirements.txt
# Core dependencies
sqlalchemy==2.0.25
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Database
psycopg2-binary==2.9.9

# Web Scraping
feedparser==6.0.11  # RSS feed parsing
newspaper3k==0.2.8  # Article extraction
beautifulsoup4==4.12.3  # HTML parsing
requests==2.31.0  # HTTP requests
lxml==5.1.0  # XML/HTML parser
python-dateutil==2.8.2  # Date parsing

# Scheduling
apscheduler==3.10.4  # Task scheduling

# NLP & Embeddings
sentence-transformers==2.7.0  # For generating embeddings (updated for compatibility)
numpy==1.26.3
scikit-learn==1.4.0  # For clustering articles into events

